memory-config-agent/requirements.txt
# Memory Config Agent dependencies
# The ncp-sdk is the only required dependency
ncp-sdk
